server/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
pyserial==3.5
pydantic==1.8.2
aiocoap==0.4.3
requests==2.26.0
click==8.0.3
rich==10.12.0
pyrtlsdr==0.2.9
